[refactor] prep for data Api endpoint * Endpoint: authentication, session handling and http clinet mgmt * WorkerEndpoint: API worker methods * DataEndpoint: TODO data Api methods
