+++ examples/node/visualize_on_image.js
@@ -1,0 +1,128 @@
+"use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
+var __asyncValues = (this && this.__asyncValues) || function (o) {
+    if (!Symbol.asyncIterator) throw new TypeError("Symbol.asyncIterator is not defined.");
+    var m = o[Symbol.asyncIterator], i;
+    return m ? m.call(o) : (o = typeof __values === "function" ? __values(o) : o[Symbol.iterator](), i = {}, verb("next"), verb("throw"), verb("return"), i[Symbol.asyncIterator] = function () { return this; }, i);
+    function verb(n) { i[n] = o[n] && function (v) { return new Promise(function (resolve, reject) { v = o[n](v), settle(resolve, reject, v.done, v.value); }); }; }
+    function settle(resolve, reject, d, v) { Promise.resolve(v).then(function(v) { resolve({ value: v, done: d }); }, reject); }
+};
+exports.__esModule = true;
+var eyepop_1 = require("../../src/eyepop");
+var canvas_1 = require("canvas");
+var openurl_1 = require("openurl");
+var node_fs_1 = require("node:fs");
+var node_path_1 = require("node:path");
+var node_os_1 = require("node:os");
+var pino_1 = require("pino");
+var process_1 = require("process");
+var logger = (0, pino_1.pino)({ level: 'info', name: 'eyepop-example' });
+var example_image_path = process_1["default"].argv[2];
+(function () { return __awaiter(void 0, void 0, void 0, function () {
+    var image, canvas, context, endpoint, results, _a, _b, result, e_1_1, tmp_dir, temp_file, buffer;
+    var e_1, _c;
+    return __generator(this, function (_d) {
+        switch (_d.label) {
+            case 0: return [4 /*yield*/, (0, canvas_1.loadImage)(example_image_path)];
+            case 1:
+                image = _d.sent();
+                canvas = (0, canvas_1.createCanvas)(image.width, image.height);
+                context = canvas.getContext("2d");
+                return [4 /*yield*/, eyepop_1.EyePop.workerEndpoint({
+                        logger: logger
+                    }).onStateChanged(function (fromState, toState) {
+                        logger.info("Endpoint changed state %s -> %s", fromState, toState);
+                    }).connect()];
+            case 2:
+                endpoint = _d.sent();
+                _d.label = 3;
+            case 3:
+                _d.trys.push([3, , 18, 20]);
+                return [4 /*yield*/, endpoint.process({ path: example_image_path })];
+            case 4:
+                results = _d.sent();
+                _d.label = 5;
+            case 5:
+                _d.trys.push([5, 11, 12, 17]);
+                return [4 /*yield*/, results];
+            case 6:
+                _a = __asyncValues.apply(void 0, [_d.sent()]);
+                _d.label = 7;
+            case 7: return [4 /*yield*/, _a.next()];
+            case 8:
+                if (!(_b = _d.sent(), !_b.done)) return [3 /*break*/, 10];
+                result = _b.value;
+                canvas.width = result.source_width;
+                canvas.height = result.source_height;
+                context.drawImage(image, 0, 0);
+                Render2d.renderer(context).draw(result);
+                _d.label = 9;
+            case 9: return [3 /*break*/, 7];
+            case 10: return [3 /*break*/, 17];
+            case 11:
+                e_1_1 = _d.sent();
+                e_1 = { error: e_1_1 };
+                return [3 /*break*/, 17];
+            case 12:
+                _d.trys.push([12, , 15, 16]);
+                if (!(_b && !_b.done && (_c = _a["return"]))) return [3 /*break*/, 14];
+                return [4 /*yield*/, _c.call(_a)];
+            case 13:
+                _d.sent();
+                _d.label = 14;
+            case 14: return [3 /*break*/, 16];
+            case 15:
+                if (e_1) throw e_1.error;
+                return [7 /*endfinally*/];
+            case 16: return [7 /*endfinally*/];
+            case 17: return [3 /*break*/, 20];
+            case 18: return [4 /*yield*/, endpoint.disconnect()];
+            case 19:
+                _d.sent();
+                return [7 /*endfinally*/];
+            case 20:
+                tmp_dir = (0, node_fs_1.mkdtempSync)((0, node_path_1.join)((0, node_os_1.tmpdir)(), 'ep-demo-'));
+                temp_file = (0, node_path_1.join)(tmp_dir, 'out.png');
+                logger.info("creating temp file: %s", temp_file);
+                buffer = canvas.toBuffer('image/png');
+                (0, node_fs_1.writeFileSync)(temp_file, buffer);
+                (0, openurl_1.open)("file://".concat(temp_file));
+                return [2 /*return*/];
+        }
+    });
+}); })();
