+++ src/eyepop/types.ts
@@ -1,11 +1,11 @@
 export interface Session {
     readonly eyepopUrl: string;
-    readonly popId: string;
     readonly accessToken: string;
     readonly validUntil: number;
 }
 
-export interface SessionPlus extends Session {
+export interface WorkerSession extends Session {
+    readonly popId: string;
     readonly baseUrl: string;
     readonly pipelineId: string;
     readonly sandboxId: string | undefined;
@@ -16,7 +16,6 @@ export enum EndpointState {
     Busy = "Busy",
     Authenticating = "Authenticating",
     FetchConfig = "FetchConfig",
-    StartingPop = "StartingPop",
     Error = "Error"
 }
