+++ tests/eyepop/endpoint.load.reconnect.test.ts
@@ -94,7 +94,7 @@ describe('EyePopSdk endpoint reconnect', () => {
             ctx.body = JSON.stringify({timestamp: fake_timestamp})
         })
 
-        const endpoint = EyePop.endpoint({
+        const endpoint = EyePop.workerEndpoint({
             eyepopUrl: server.getURL().toString(),
             auth: {secretKey: test_secret_key},
             popId: test_pop_id,
