+++ tests/eyepop/endpoint.upload.test.ts
@@ -73,7 +73,7 @@ describe('EyePopSdk endpoint module upload', () => {
             ctx.body = JSON.stringify({timestamp: fake_timestamp})
         })
 
-        const endpoint = EyePop.endpoint({
+        const endpoint = EyePop.workerEndpoint({
             eyepopUrl: server.getURL().toString(),
             auth: {secretKey: test_secret_key},
             popId: test_pop_id,
@@ -105,7 +105,7 @@ describe('EyePopSdk endpoint module upload', () => {
         const image_path = './tests/does_not_exist.dummy'
         const {authenticationRoute, popConfigRoute} = prepMockServer(server, test_pop_id, test_pipeline_id)
 
-        const endpoint = EyePop.endpoint({
+        const endpoint = EyePop.workerEndpoint({
             eyepopUrl: server.getURL().toString(),
             auth: {secretKey: test_secret_key},
             popId: test_pop_id,
